feat : 유저 관심장소 관련 기능

+++ src/main/java/com/example/ReviewZIP/domain/store/Stores.java
@@ -27,10 +27,10 @@ public class Stores extends BaseEntity {
     private String address;
 
     @Column(nullable = false)
-    private BigDecimal longitude;
+    private Double longitude;
 
     @Column(nullable = false)
-    private BigDecimal latitude;
+    private Double latitude;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id")
